src/workflow_endpoint/requirements.txt
Flask==1.0.2
Flask-API==1.1
Flask-Cors==3.0.8
requests==2.22.0
python-dotenv==0.10.2
# simplejson==3.16.0
mysql-connector-python==8.0.16